css/css-backgrounds/parsing/background-position-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-position with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-position">
<meta name="assert" content="background-position supports the full grammar '<bg-position>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("background-position", "1px", ["1px", "1px center"]);
test_valid_value("background-position", "1px center", ["1px", "1px center"]);
test_valid_value("background-position", "-2% -3%");
test_valid_value("background-position", "5% top");
test_valid_value("background-position", "center", ["center", "center center"]);
test_valid_value("background-position", "center center", ["center", "center center"]);
test_valid_value("background-position", "center 6px");
test_valid_value("background-position", "center left", "left center");
test_valid_value("background-position", "center right 7%", "right 7% center");
test_valid_value("background-position", "center bottom");
test_valid_value("background-position", "center top 8px");
test_valid_value("background-position", "left", ["left", "left center"]);
test_valid_value("background-position", "right 9%");
test_valid_value("background-position", "left 10px center");
test_valid_value("background-position", "right 11% bottom");
test_valid_value("background-position", "left 12px top 13px");
test_valid_value("background-position", "right center");
test_valid_value("background-position", "left bottom");
test_valid_value("background-position", "right top 14%");
test_valid_value("background-position", "bottom", "center bottom");
test_valid_value("background-position", "top 15px center", "center top 15px");
test_valid_value("background-position", "bottom 16% left", "left bottom 16%");
test_valid_value("background-position", "top 17px right 18px", "right 18px top 17px");
test_valid_value("background-position", "bottom center", "center bottom");
test_valid_value("background-position", "top left", "left top");
test_valid_value("background-position", "bottom right 19%", "right 19% bottom");
</script>
</body>
</html>
